vsplunk: show query history, reload sheet for new results

+++ vsplunk/vsplunk.py
@@ -1,13 +1,90 @@
 #!/usr/bin/env python3
+"""
+TODO
+- SplunkSheet dive row, check for existing sheet and reload
+- Save SplunkSheet as .spl file
+- Open .spl file
+"""
+import re
 import argparse
 import configparser
 import os
-from pathlib import Path
+import datetime
 
-from visidata import asyncthread, ColumnItem, Sheet, deduceType, vd, run, Progress
+from visidata import *
+
 
 __version__ = 0.1
 __author__ = 'Lucas Messenger'
+
+
+class SplunkSheet(Sheet):
+    def __init__(self, name, **kwargs):
+        self.filetype = 'spl'
+        self.rowtype = 'queries'
+        self.rows = []
+        self.columns = [
+            ColumnItem('query', 0, type=str, width=20),
+            ColumnItem('last_run', 1,
+                       fmtstr='%Y-%m-%d %H:%M:%S',
+                       type=date, width=20)
+        ]
+
+        super().__init__(name, **kwargs)
+        self.name = name
+
+
+class SplunkSearchSheet(Sheet):
+    def __init__(self, query, **kwargs):
+        name = re.sub('[^a-zA-Z0-9]', '', query)
+        self.rowtype = 'results'
+        self.colnames = {}
+        self.columns = []
+
+        super().__init__(name, **kwargs)
+        self.query = query
+        self.reload()
+
+    def addRow(self, row, index=None):
+        super().addRow(row, index=index)
+        if isinstance(row, dict):
+            for k, v in row.items():
+                if k not in self.colnames:
+                    if k.startswith('_') and not k == '_time':
+                        c = ColumnItem(k, type=deduceType(v), width=0)
+                    else:
+                        c = ColumnItem(k, type=deduceType(v))
+                    self.colnames[k] = c
+                    self.addColumn(c)
+            return row
+
+    @asyncthread
+    def reload(self):
+        """
+        add rows from query results
+        """
+        self.colnames = {}
+        self.columns = []
+        self.rows = []
+
+        with Progress(total=0, gerund='splunking'):
+            for row in self.search():
+                self.addRow(row)
+
+    def search(self):
+        import splunklib.results
+        search_settings = {'search_mode': 'normal',
+                           'count': 0}
+
+        if self.query.startswith('search'):
+            resp = splunkc.jobs.export(self.query, **search_settings)
+        else:
+            resp = splunkc.jobs.export('search {}'.format(self.query),
+                                       **search_settings)
+
+        for result in splunklib.results.ResultsReader(resp):
+            if isinstance(result, dict):
+                yield result
 
 
 def get_args():
@@ -15,7 +92,9 @@ def get_args():
     get args
     """
     parser = argparse.ArgumentParser()
-    parser.add_argument('-c', '--config', default=os.path.join(Path.home(), '.vsplunk'))
+    parser.add_argument('-c', '--config',
+                        help='vsplunk configuration file',
+                        default=Path('~/.vsplunk'))
     return parser.parse_args()
 
 
@@ -28,79 +107,46 @@ def read_config(path):
 
     if 'SPLUNK' in config.sections():
         return config['SPLUNK']
+
     config['SPLUNK'] = {'host': 'localhost',
                         'port': 8089,
                         'scheme': 'https',
                         'username': 'admin',
                         'password': 'changeme'}
 
-    with open(path, 'w') as fp:
-        config.write(fp)
-    return read_config(path)
+    if not os.path.exists(path):
+        with open(path, 'w') as fp:
+            config.write(fp)
+        return read_config(path)
+    vd.status('unable to read ~/.vsplunk', priority=3)
+    return config['SPLUNK']
 
 
-class SplunkSheet(Sheet):
-    def __init__(self, name, **kwargs):
-        super().__init__(name, **kwargs)
-        self.colnames = {}
-        self.columns = []
-        self.rows = []
-        try:
-            import splunklib.client
-            self.client = splunklib.client.connect(**kwargs)
-            vd.status('connected to splunk', priority=2)
-            vd.status('issue query to begin', priority=-1)
-        except Exception as e:
-            vd.status(e, priority=3)
+def search_splunk():
+    query = vd().input('splunk-query: ', 'splunk')
+    vd.splunk.addRow((query, datetime.datetime.utcnow()), index=None)
+    return vd.push(SplunkSearchSheet(query))
 
-    def addRow(self, row, index=None):
-        super().addRow(row, index=index)
-        if isinstance(row, dict):
-            for k, v in row.items():
-                if k not in self.colnames:
-                    if k.startswith('_'):
-                        c = ColumnItem(k, type=deduceType(v), width=0)
-                    else:
-                        c = ColumnItem(k, type=deduceType(v))
-                    self.colnames[k] = c
-                    self.addColumn(c)
 
-    @asyncthread
-    def runQuery(self, query):
-        """
-        add rows from query results
-        """
-        # clear the board on new query
-        self.colnames = {}
-        self.columns = []
-        self.rows = []
-
-        with Progress(total=0, gerund='splunking'):
-            for row in self.search(query):
-                self.addRow(row)
-
-    def search(self, query):
-        import splunklib.results
-        search_settings = {'search_mode': 'normal',
-                           'count': 0}
-
-        if query.startswith('search'):
-            resp = self.client.jobs.export(query, **search_settings)
-        else:
-            resp = self.client.jobs.export('search {}'.format(query),
-                                           **search_settings)
-
-        for result in splunklib.results.ResultsReader(resp):
-            if isinstance(result, dict):
-                yield result
-
-Sheet.addCommand('^N', 'splunk-query', 'runQuery(input("query: ", "splunk"))')
+addGlobals(globals())
+Sheet.addCommand('^N', 'splunk-query', 'search_splunk()')
+SplunkSheet.addCommand(ENTER, 'dive-row', 'vd.push(SplunkSearchSheet(cursorRow[0]))', 'search Splunk with query'),
+vd.splunk = SplunkSheet('vsplunk')
 
 
 def main_vsplunk():
     args = get_args()
     config = read_config(args.config)
-    run(SplunkSheet(name='vsplunk', **config))
+    try:
+        import splunklib.client
+        global splunkc
+        splunkc = splunklib.client.connect(**config)
+        vd.status('connected to splunk', priority=2)
+        vd.status('issue query to begin', priority=-1)
+    except Exception as e:
+        vd.status(e, priority=3)
+
+    run(vd.splunk)
 
 
 if __name__ == '__main__':
